% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_typo.R
\name{check_typo}
\alias{check_typo}
\title{Check for typo in your list of species names}
\usage{
check_typo(species, to_return = FALSE, suppress_warning = FALSE)
}
\arguments{
\item{species}{A character vector}

\item{to_return}{If TRUE the function to return the array of suggestions}

\item{suppress_warning}{Logical, if \code{TRUE} the function will not display a warning message}
}
\value{
If all names are correct, the function returns a message indicating that you can procced to the next step
In any misspelled names are found, a warning is displayed. The warning message, highlights all
names and provides three suggestions for each one.
}
\description{
Check for typo in your list of species names
}
\examples{
# It looks like this, when there are no mistakes
species_list <- c( "Sternbergia colchiciflora Waldst. et Kit.","Scorzonera humilis L.",
"Ceratophyllum demersum L.","Lathyrus sylvestris L.","Sedum sexangulare L.")
check_typo(species_list)

# Now let's take the same species_list as above, but let's introduce some mistakes. I am gonna
#change the first and last names.
species_list <- c( "Stenrbegria colchiciflra Waldst. et Kit.","Scorzonera humilis L.",
"Ceratophyllum demersum L.","Lathyrus sylvestris L.","Sedum sektangulare")
check_typo(species_list)
}
