% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_typo.R
\name{check_typo_trait}
\alias{check_typo_trait}
\title{Spell checks the traits}
\usage{
check_typo_trait(traits, verbose = TRUE)
}
\arguments{
\item{traits}{A list of character, the trait names}

\item{verbose}{If TRUE you'll get any message when run. It is set to FALSE when is called by
another function}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Spell checks the traits
}
