% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert a standard scientific name to the ones found in PADAPT}
\usage{
convert(list)
}
\arguments{
\item{list}{Your list of species}
}
\value{
A list. By calling \verb{$correct}, you can inspect all the correct names, and by calling
\verb{$errors} you will see all the names which is neither in our database, nor a synonym for one.
}
\description{
A standard scientific name is usually made up from two words. In PADAPT, the species column also
contains the name of researcher who first described it. So this function, \code{convert()} takes
your list of standard names, applies the \code{safe_search()} function to find the your species
which are also within PADAPT. It also saves all the wrong ones, and by running \code{get_synonyms()}
it tries to find a synonym from GBIF which might be found in PADAPT also.
}
