% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_means}
\alias{get_means}
\title{Get means of traits which are collected from multiple sources}
\usage{
get_means(df, traits)
}
\arguments{
\item{df}{Data already selected from the database, look up \code{padapt_query()} to understand more}

\item{traits}{The list of traits requested by the used in \code{padapt_query()}}
}
\value{
A data frame which contains only the means of the specific traits
}
\description{
In this dataset all those traits which are collected from multiple sources are marked with
a number, such as SLA1, SLA2, SLA3.... . This makes it easy to distinguish all the traits
for which we need to calculate the mean value. This is achieved by \code{grep()}, and to collapse
the names \code{gsub()} is the chosen tool. The mean itself is calculated by \code{rowMeans()}.
}
