% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_suggestions}
\alias{get_suggestions}
\title{Gives suggestions based on the misstyped species names}
\usage{
get_suggestions(names)
}
\arguments{
\item{names}{An array with the misstyped names}
}
\value{
A data frame, built from the ID-s of the recommended species.
With there ID-s we can extract the name of the correct species names.
}
\description{
It uses the \code{stringdist::stringsim()} function for computing similarity scores. By ordering
these scores we select three options with the highest scores for each of the species
}
