% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{get_synonyms}
\alias{get_synonyms}
\title{Get synonyms from GBIF}
\usage{
get_synonyms(x)
}
\arguments{
\item{x}{The list of species you're trying to get a synonym for}
}
\value{
A list which contains the proper synonyms, which are in PADAPT, and the names of the
species that are still missing
}
\description{
Get synonyms from GBIF
}
