% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{get_synonyms_backbone}
\alias{get_synonyms_backbone}
\title{Get synonyms from GBIF with name_backbone()}
\usage{
get_synonyms_backbone(x)
}
\arguments{
\item{x}{The list of species you're trying to get a synonym for}
}
\value{
A list which contains the proper synonyms, which are in PADAPT, and the names of the
species that are still missing
}
\description{
This function is based on \code{\link[rgbif:name_backbone]{rgbif::name_backbone()}}
}
