% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{get_synonyms_lookup}
\alias{get_synonyms_lookup}
\title{Searching for synonyms with name_lookup()}
\usage{
get_synonyms_lookup(x)
}
\arguments{
\item{x}{Names of species you are trying to get synonyms for}
}
\value{
A list of four elements, these are:
\describe{
\item{syn}{All synonyms that are also present in the PADAPT database.}
\item{errors}{Input names for which no results were returned from GBIF.}
\item{correct_input}{Input names for which at least one plausible synonym was found in PADAPT.}
\item{wrong_input}{Input names for which synonyms were found, but none were present in PADAPT.}
}
}
\description{
This function is based on \code{\link[rgbif:name_lookup]{rgbif::name_lookup()}}
}
