% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padapt.R
\docType{data}
\name{padapt}
\alias{padapt}
\title{Pannonian Databse Of Plant Traits}
\format{
A data frame with 2745 rows and 79 columns.

NOTE: This version also includes a column for the genus, which is not found in the original.
To read about the variable, I recommend their official website: https://padapt.eu/info.html
}
\source{
\url{https://figshare.com/articles/dataset/PADAPT_1_0_the_Pannonian_Database_of_Plant_Traits/21937157?file=41684067}
}
\usage{
padapt
}
\description{
You can read more about this database on the following website: https://padapt.eu/en
}
\author{
Sonkoly et. al.
}
\keyword{datasets}
