% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padapt_query.R
\name{padapt_query}
\alias{padapt_query}
\title{Queries the database}
\usage{
padapt_query(list_of_species, list_of_traits)
}
\arguments{
\item{list_of_species}{A list which contains the names of the species}

\item{list_of_traits}{A list which contains the names of the traits}
}
\value{
The subset of the database you are interested in.
}
\description{
Queries the database
}
\examples{

#First, we deifne a list of species, and a list of traits
species = c("Stipa capillata L.", "Abies alba Mill." )
traits = c("sla1","sla2","sla3")
padapt_query(species, traits)
}
