% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_species_traits.R
\name{rand_species}
\alias{rand_species}
\title{Get random number of species from the table}
\usage{
rand_species(n)
}
\arguments{
\item{n}{The number of species you want to get}
}
\value{
A character vector
}
\description{
Get random number of species from the table
}
\examples{
rand_species(10)
}
