% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_species_traits.R
\name{rand_traits}
\alias{rand_traits}
\title{Get random number of traits from the table}
\usage{
rand_traits(n)
}
\arguments{
\item{n}{The number of traits you want to get}
}
\value{
A character vector
}
\description{
Get random number of traits from the table
}
\examples{
rand_species(10)
}
