% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{safe_search}
\alias{safe_search}
\title{A search method built upon tryCatch}
\usage{
safe_search(x)
}
\arguments{
\item{x}{A list of names}
}
\value{
A list of 2 parts. The first part collects all the proper names, already found in the
database, while the second part collects all those missing.
}
\description{
A search method built upon tryCatch
}
