% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_species.R
\name{search_species}
\alias{search_species}
\title{Tool for searching species names}
\usage{
search_species(name)
}
\arguments{
\item{name}{A species name of a character string contained within the species name.}
}
\value{
A character vector of species names containing the search string.
}
\description{
Tool for searching species names
}
\examples{
search_species('Stipa')

# In case you are not sure about the whole name you can input even characters
search_species('albu')

}
