% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_typo.R
\name{check_typo}
\alias{check_typo}
\title{Check if you've made a typo in your list of species names}
\usage{
check_typo(species)
}
\arguments{
\item{species}{The list of names}
}
\value{
A message if the names are correct and a warning if there is a mistake. Also returns the wrong names.
}
\description{
Check if you've made a typo in your list of species names
}
\examples{
# generate a random list of species first
species_list <- c( "Sternbergia colchiciflora Waldst. et Kit.","Scorzonera humilis L.",
"Ceratophyllum demersum L.","Lathyrus sylvestris L.","Sedum sexangulare L.")
check_typo(species_list)
}
