% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_typo.R
\name{check_typo}
\alias{check_typo}
\title{Check if you've made a typo in your list of species names}
\usage{
check_typo(species, to_return = FALSE, suppress_warning = FALSE)
}
\arguments{
\item{species}{A vector of characters}

\item{to_return}{TRUE in case you want the function to return the array of recommendation}

\item{suppress_warning}{Logical, if \code{TRUE} the function will not pop up it's warning message}
}
\value{
If everything is correct, you'll get a message which means you can go ahead to next step.
In case of having something wrong, a warning pops up. In the warning message, all the misstyped
names will be highlighted, and there will be 3 recommendation for each and every one
of the names
}
\description{
Check if you've made a typo in your list of species names
}
\examples{
# It looks like this, when there are no mistakes
species_list <- c( "Sternbergia colchiciflora Waldst. et Kit.","Scorzonera humilis L.",
"Ceratophyllum demersum L.","Lathyrus sylvestris L.","Sedum sexangulare L.")
check_typo(species_list)

# Now let's take the same species_list as above, but let's introduce some mistakes. I am gonna
#change the first and last names.
species_list <- c( "Stenrbegria colchiciflra Waldst. et Kit.","Scorzonera humilis L.",
"Ceratophyllum demersum L.","Lathyrus sylvestris L.","Sedum sektangulare")
check_typo(species_list)
}
