% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Covert a standard scientific name to the ones found in PADAPT}
\usage{
convert(list)
}
\arguments{
\item{list}{Your list of species}
}
\value{
Characters, a list of the species names
}
\description{
A standard scientific name is usually made up from two words. In PADAPT, a species column also
contains the name of researcher who first described it. So this function, \code{convert()} takes
your list of standard names, applies the \code{check_typo()} function and returns a list of correct
names which you can use instantly for the query.
}
