% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert a standard scientific name to PADAPT-compatible names}
\usage{
convert(list, table = FALSE)
}
\arguments{
\item{list}{A character vector of species names.}

\item{table}{Logical; if \code{TRUE}, the function also returns a data frame showing the conversion results.
Default is \code{FALSE}.}
}
\value{
A named list with the following elements:
\describe{
\item{correct}{A character vector of species names successfully matched to
PADAPT, including author citations.}
\item{errors}{A character vector of input names for which neither the original
name nor any GBIF-derived synonym could be matched to the PADAPT database.}
\item{df}{A data frame showing you how the inputs were converted. This is returned only
if you specify \code{table = TRUE}}
}
}
\description{
Scientific names are typically composed of two parts. In PADAPT however, the species column also
includes the name of the researcher who first described the species. The \code{convert()} function takes
a vector of species names and attempts to match them to the corresponding names used in PADAPT.

First, it applies \code{\link[=safe_search]{safe_search()}} to identify names that already match entries in PADAPT. For
names that are not found, it attempts to retrieve accepted synonyms for GBIF using \code{\link[=get_synonyms_backbone]{get_synonyms_backbone()}}
and \code{\link[=get_synonyms_lookup]{get_synonyms_lookup()}}. These candidates are then checked again against PADAPT to find valid
mathces.
}
\seealso{
\code{\link[=safe_search]{safe_search()}}, \code{\link[=get_synonyms_backbone]{get_synonyms_backbone()}}, \code{\link[=get_synonyms_lookup]{get_synonyms_lookup()}}
}
