% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert a standard scientific name to PADAPT compatible names}
\usage{
convert(list, table = FALSE)
}
\arguments{
\item{list}{A character vector of species names.}

\item{table}{A boolean operator which defines what the function returns, \code{FALSE} by default.}
}
\value{
A named list with the following elements:
\describe{
\item{correct}{A character vector of species names successfully matched to
PADAPT, including author citations.}
\item{errors}{A character vector of input names for which neither the original
name nor any GBIF-derived synonym could be matched to the PADAPT database.}
\item{df}{A dataframe in which you may inspect what your inputs have been converted into. The
function only returns df if you specify \code{table = TRUE}}
}
}
\description{
Scientific names are typically made up from two parts. In PADAPT however, the species column also
contains the name of researcher who first described it. The \code{convert()} function takes
a vector of species names and tries to match them to the corresponding names used in PADAPT.

First, it applies the \code{\link[=safe_search]{safe_search()}} to identify names that already match entries in PADAPT. For
names that are not found it attempts to retrieve accepted synonyms for GBIF using \code{\link[=get_synonyms_backbone]{get_synonyms_backbone()}}
and \code{\link[=get_synonyms_lookup]{get_synonyms_lookup()}}. These candidates then are checked again against PADAPT to find valid
mathces.
}
\seealso{
\code{\link[=safe_search]{safe_search()}}, \code{\link[=get_synonyms_backbone]{get_synonyms_backbone()}}, \code{\link[=get_synonyms_lookup]{get_synonyms_lookup()}}
}
