% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padapt_query.R
\name{padapt_query}
\alias{padapt_query}
\title{Search the database}
\usage{
padapt_query(
  list_of_species,
  list_of_traits = NULL,
  preset = "all",
  means = FALSE,
  just_means = FALSE
)
}
\arguments{
\item{list_of_species}{A list which contains the names of the species}

\item{list_of_traits}{A list which contains the names of the traits}

\item{means}{Boolean operator. \code{FALSE} by default, if \code{TRUE} the mean will be calculated for
traits with multiple sources, for example the leaf traits. There is 5 options to choose from:
\verb{'all','none','hab_str','repr','kary','dist_cons','ecol_ind','leaf_traits'}. The traits has
been categorized by this table: \url{https://padapt.eu/info.html}}

\item{just_means}{Boolean operator. \code{FALSE} by default, if \code{TRUE} only the mean trait values
will be included in the results in case of a trait having multiple sources}
}
\value{
The subset of the database you are interested in.
}
\description{
Searches the database by the species you are interested in, extracting all the traits you
might need for your analysis.
}
\examples{

#First, we deifne a list of species, and a list of traits
species = c("Acorus calamus L.", "Ambrosia artemisiifolia L." )
traits = c("sla1","sla2","sla3","sla4")
padapt_query(species, traits)
}
