% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_species.R
\name{search_species}
\alias{search_species}
\title{Tool for searching species names}
\usage{
search_species(name)
}
\arguments{
\item{name}{The name of the species or a subset of character which are in the name}
}
\value{
An array of names which contains the search characters of word.
}
\description{
Tool for searching species names
}
\examples{
search_species('Stipa')

# In case you are not sure about the whole name you can input even characters
search_species('albu')

}
